Updated Testing Logic
